+++ arrow-core-data/src/main/kotlin/arrow/core/MonadContinuation.kt
@@ -2,7 +2,6 @@ package arrow.core
 
 import arrow.Kind
 import arrow.typeclasses.suspended.BindSyntax
-import arrow.typeclasses.suspended.Invoke
 import kotlin.coroutines.Continuation
 import kotlin.coroutines.CoroutineContext
 import kotlin.coroutines.EmptyCoroutineContext
@@ -11,9 +10,6 @@ import kotlin.coroutines.intrinsics.startCoroutineUninterceptedOrReturn
 
 @RestrictsSuspension
 interface EagerBind<F> : BindSyntax<F>
-
-@RestrictsSuspension
-interface EagerInvoke<F> : Invoke<F>
 
 class ShortCircuit(val value: Any?) : RuntimeException(null, null) {
   override fun fillInStackTrace(): Throwable = this
