--- arrow-core-data/src/main/kotlin/arrow/core/computations/eitherContinuation.kt
@@ -1,38 +1,0 @@
-package arrow.core.computations
-
-import arrow.Kind
-import arrow.continuations.generic.DelimContScope
-import arrow.core.EagerInvoke
-import arrow.core.Either
-import arrow.core.EitherPartialOf
-import arrow.core.fix
-import arrow.typeclasses.suspended.Invoke
-
-object eitherContinuation {
-
-  fun <E, A> eager(c: suspend EagerInvoke<EitherPartialOf<E>>.() -> A): Either<E, A> =
-    DelimContScope.reset {
-      Either.Right(
-        c(object : EagerInvoke<EitherPartialOf<E>> {
-          override suspend fun <A> Kind<EitherPartialOf<E>, A>.invoke(): A =
-            when (val v = fix()) {
-              is Either.Right -> v.b
-              is Either.Left -> shift { v }
-            }
-        })
-      )
-    }
-
-  suspend operator fun <E, A> invoke(c: suspend Invoke<EitherPartialOf<E>>.() -> A): Either<E, A> =
-    DelimContScope.reset {
-      Either.Right(
-        c(object : Invoke<EitherPartialOf<E>> {
-          override suspend fun <A> Kind<EitherPartialOf<E>, A>.invoke(): A =
-            when (val v = fix()) {
-              is Either.Right -> v.b
-              is Either.Left -> shift { v }
-            }
-        })
-      )
-    }
-}
