--- arrow-core-data/src/main/kotlin/arrow/core/computations/validatedContinuations.kt
@@ -1,39 +1,0 @@
-package arrow.core.computations
-
-import arrow.Kind
-import arrow.continuations.generic.DelimContScope
-import arrow.core.EagerInvoke
-import arrow.core.Valid
-import arrow.core.Validated
-import arrow.core.ValidatedPartialOf
-import arrow.core.fix
-import arrow.typeclasses.suspended.Invoke
-
-object validatedContinuations {
-
-  fun <E, A> eager(c: suspend EagerInvoke<ValidatedPartialOf<E>>.() -> A): Validated<E, A> =
-    DelimContScope.reset {
-      Valid(
-        c(object : EagerInvoke<ValidatedPartialOf<E>> {
-          override suspend fun <A> Kind<ValidatedPartialOf<E>, A>.invoke(): A =
-            when (val v = fix()) {
-              is Validated.Valid -> v.a
-              is Validated.Invalid -> shift { v }
-            }
-        })
-      )
-    }
-
-  suspend operator fun <E, A> invoke(c: suspend Invoke<ValidatedPartialOf<E>>.() -> A): Validated<E, A> =
-    DelimContScope.reset {
-      Valid(
-        c(object : Invoke<ValidatedPartialOf<E>> {
-          override suspend fun <A> Kind<ValidatedPartialOf<E>, A>.invoke(): A =
-            when (val v = fix()) {
-              is Validated.Valid -> v.a
-              is Validated.Invalid -> shift { v }
-            }
-        })
-      )
-    }
-}
