+++ arrow-core-data/src/main/kotlin/arrow/typeclasses/MonadContinuations.kt
@@ -31,10 +31,13 @@ open class MonadContinuation<F, A>(
   open fun returnedMonad(): Kind<F, A> = returnedMonad
 
   // TODO: Deprecate
-  override suspend fun <B> Kind<F, B>.bind(): B =
+  override suspend fun <A> Kind<F, A>.bind(): A =
+    invoke()
+
+  override suspend fun <A> Kind<F, A>.invoke(): A =
     suspendCoroutineUninterceptedOrReturn { c ->
       val labelHere = c.stateStack // save the whole coroutine stack labels
-      returnedMonad = this.flatMap { x: B ->
+      returnedMonad = this.flatMap { x: A ->
         c.stateStack = labelHere
         c.resume(x)
         returnedMonad
